refactor :: brand-product 연관관계 제거에 따른 변경
